device_simulator/requirements.txt
paho-mqtt==2.1.0
dotenv==0.21.0